+++ components/[pageId]/DocumentPage/DocumentPage.tsx
@@ -16,6 +16,7 @@ import { blockLoad, databaseViewsLoad } from 'components/common/BoardEditor/foca
 import { useAppDispatch, useAppSelector } from 'components/common/BoardEditor/focalboard/src/store/hooks';
 import { CharmEditor } from 'components/common/CharmEditor';
 import { CardPropertiesWrapper } from 'components/common/CharmEditor/CardPropertiesWrapper';
+import { handleImageFileDrop } from 'components/common/CharmEditor/components/@bangle.dev/base-components/image';
 import type { FrontendParticipant } from 'components/common/CharmEditor/components/fiduswriter/collab';
 import type { ConnectionEvent } from 'components/common/CharmEditor/components/fiduswriter/ws';
 import { SnapshotVoteDetails } from 'components/common/CharmEditor/components/inlineVote/components/SnapshotVoteDetails';
@@ -26,7 +27,6 @@ import { useBountyPermissions } from 'hooks/useBountyPermissions';
 import { useCharmEditor } from 'hooks/useCharmEditor';
 import { usePageSidebar } from 'hooks/usePageSidebar';
 import { useVotes } from 'hooks/useVotes';
-import { getFileData } from 'lib/file/getFileData';
 import type { PageWithContent } from 'lib/pages/interfaces';
 import type { PageContent } from 'lib/prosemirror/interfaces';
 import { fontClassName } from 'theme/fonts';
@@ -215,34 +215,11 @@ function DocumentPage({ page, refreshPage, savePage, readOnly = false, close }: 
               flexDirection: 'column'
             }}
             ref={containerRef}
-            onDrop={(event) => {
-              // prevent drop event from firing when dropping into the editor
-              if (
-                event.dataTransfer == null ||
-                readOnly ||
-                (event.target instanceof HTMLElement && event.target.parentElement?.id !== 'document-scroll-container')
-              ) {
-                return;
-              }
-              const files = getFileData(event.dataTransfer, 'image/*', true);
-
-              if (!files || files.length === 0) {
-                return;
-              }
-              event.preventDefault();
-
-              const imageFileDropEvent = new CustomEvent('imageFileDrop', {
-                detail: {
-                  files
-                }
-              });
-
-              const bangleEditorCoreElement = document.querySelector(`.bangle-editor-core`);
-
-              if (bangleEditorCoreElement) {
-                bangleEditorCoreElement.dispatchEvent(imageFileDropEvent as Event);
-              }
-            }}
+            onDrop={handleImageFileDrop({
+              pageId: page.id,
+              readOnly,
+              parentElementId: 'document-scroll-container'
+            })}
           >
             <PageTemplateBanner parentId={page.parentId} pageType={page.type} />
             {/* temporary? disable editing of page meta data when in suggestion mode */}
